+++ blueprints/codes/eurocode/en_1993_1_1_2022/chapter_8_ultimate_limit_state/formula_8_22.py
@@ -1,6 +1,6 @@
-"""Formula 8.22 from EN 1993-1-1:2025: Chapter 8 - Ultimate Limit State."""
+"""Formula 8.22 from EN 1993-1-1:2022: Chapter 8 - Ultimate Limit State."""
 
-from blueprints.codes.eurocode.en_1993_1_1_2025 import EN_1993_1_1_2025
+from blueprints.codes.eurocode.en_1993_1_1_2022 import EN_1993_1_1_2022
 from blueprints.codes.formula import Formula
 from blueprints.codes.latex_formula import LatexFormula, latex_replace_symbols
 from blueprints.type_alias import N
@@ -11,7 +11,7 @@ class Form8Dot22CheckShearForce(Formula):
     r"""Class representing formula 8.22 for checking the design value of the shear force."""
 
     label = "8.22"
-    source_document = EN_1993_1_1_2025
+    source_document = EN_1993_1_1_2022
 
     def __init__(
         self,
@@ -20,7 +20,7 @@ class Form8Dot22CheckShearForce(Formula):
     ) -> None:
         r"""Check the design value of the shear force at each cross section.
 
-        EN 1993-1-1:2025 art.8.2.6(1) - Formula (8.22)
+        EN 1993-1-1:2022 art.8.2.6(1) - Formula (8.22)
 
         Parameters
         ----------
